third_party/WebKit/LayoutTests/fast/dom/htmlallcollection-enumerated-properties-expected.txt

This tests verifies the enumerated properties on HTMLAllCollection and their order.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS htmlAllCollection.__proto__ is HTMLAllCollection.prototype
PASS htmlAllCollection.__proto__.__proto__ is Object.prototype
PASS htmlAllCollection.length is 34
PASS enumeratedProperties is expectedEnumeratedProperties
PASS successfullyParsed is true

TEST COMPLETE
 
  
   
